feat(平移): 完成移动操作（左移，下移，右移，上移）

+++ CGLineSegment.cpp
@@ -38,7 +38,7 @@ bool CGLineSegment::Render(CGRenderContext* pRC, CGCamera* pCamera)
 	return true;
 }
 
-void CGLineSegment::Translate(double tx, double ty)
+void CGLineSegment::Translate(float tx, float ty)
 {
 	mStart.x += tx;
 	mStart.y += ty;
